# Faster Rationals by avoiding unnecessary divgcd