This PR prevents `divgcd` from being unnecessarily computed when creating a `Rational` with numerators and denominators that are known to be coprime. This is entirely non-breaking.

Addresses #11522

## The history

> I think if we really wanted to tweak performance of rationals, one way would be to have an option to not perform the gcd in the constructor when it isn't required, e.g. for Rational(::Int) or after a multiplication or division. (@simonbyrne)
> That seems like it would be an improvement. I guess that would require having a way of constructing Rational values, bypassing the gcd. (@StefanKarpinski)

(#8672)

## The implementation

~I add two constructors to `Rational` that take the numerator, the denominator, and either `Val(false)` or `Val(true)`. The original constructor is left unchanged (that would be breaking).~

~The two new constructors do not reduce the fraction to its irreducible form. Since the `Rational` infrastructure relies on that fact however, using any of these two new constructors to create an ill-formed rational (one whose numerator and denominator are not coprime) results in undefined behaviour. I did not document the new constructors for this very reason, but if you believe they should be exported alongside the warning I can add some documentation.~

~The difference between the two constructors is that the one with `Val(true)` performs no check whatsoever, whereas the one with `Val(false)` still checks whether the denominator is negative and errors if it is equal to `typemin(T)` (to conform to #32569).~

~EDIT: see discussion below. Intermediate version with `unsafe_rational` at https://github.com/JuliaLang/julia/pull/35492#issuecomment-615132277. Current version with keyword arguments at https://github.com/JuliaLang/julia/pull/35492#issuecomment-617939475. Other propositions: https://github.com/JuliaLang/julia/pull/35492#issuecomment-617922722 and https://github.com/JuliaLang/julia/pull/35492#issuecomment-617997997~

EDIT: See current state of the PR at https://github.com/JuliaLang/julia/pull/35492#issuecomment-622527564. This PR is non-breaking and does not add any new export.

The new constructor is used to remove some unnecessary checks in the existing code of `rational.jl`.


The idea for the implementation comes from @JeffreySarnoff's [FastRationals.jl.](https://github.com/JeffreySarnoff/FastRationals.jl) But this PR does not provide any additional speed from relaxing the constraint that the numerator and denominator should be coprime.

## The silly benchmark

EDIT: For more serious benchmarking work, see https://github.com/JuliaLang/julia/pull/35492#issuecomment-623480983

I am using the exact same function as in #35444. This shows the new improvement compared to current master (on which #35444 has been merged).
```julia
julia> function foo(::Type{T}) where T<:Integer
           x = Rational{T}(1, 100)
           n = T(10000)
           k = one(T)
           return maximum(i-((i+((x+i)-i))-x) for i in k:n)
       end
foo (generic function with 3 methods)

julia> @btime foo(Int);
  1.147 ms (0 allocations: 0 bytes)   # Currently
  442.850 μs (0 allocations: 0 bytes) # This PR

julia> @btime foo(UInt);
  1.113 ms (0 allocations: 0 bytes)   # Currently
  311.702 μs (0 allocations: 0 bytes) # This PR

julia> @btime foo(BigInt);
  20.428 ms (800012 allocations: 15.56 MiB) # Currently
  12.896 ms (540012 allocations: 10.38 MiB) # This PR
```

## The bonus

This PR also incidentally fixes (what I believe to be) two bugs:
- `rationalize(T, x)` with `T<:Unsigned` and `x<0` used to yield `one(T)//zero(T)` aka infinity. It now errors with `OverflowError: cannot negate unsigned number`.
- `typemin(Rational{T})` with `T<:Unsigned` used to be `one(T)//zero(T)` (infinity again), it is now `zero(T)//one(T)`.