@Liozou 2020-04-15
I forgot to mention that the idea for the implementation comes from @JeffreySarnoff's [FastRationals.jl](https://github.com/JeffreySarnoff/FastRationals.jl). But this PR does not provide any additional speed from relaxing the constraint that the numerator and denominator should be coprime.