@JeffreySarnoff 2020-04-16
What do you find when benchmarking this relative to the current implementation?