@KristofferC 2020-04-16
I don't think those actually cause a failure (just scary print out)