@Liozou 2020-04-16
Ah yes you're right. The failures look unrelated to the PR (download issues it seems) and CI passed yesterday so I'll rerun them to check.