@Liozou 2020-04-17
Update on the constructors: the new one and only inner constructor for `Rational` is `unsafe_rational` now, which optionally takes a type parameter like the original constructor. It is documented (but not exported, that would require a triage call), here is the documentation:

>     unsafe_rational(num::Integer, den::Integer; checksign=false)
>     unsafe_rational{T<:Integer}(num::Integer, den::Integer; checksign::Bool=false)
> 
> Create a `Rational` with numerator `num` and denominator `den`.
> 
> The `unsafe` prefix on this function indicates that no check is performed on
> `num` and `den` to ensure that the resulting `Rational` is well-formed.
> A `Rational` is well-formed if its numerator and denominator are coprime and if
> the denominator is non-negative.
> 
> `checksign` optionally specifies whether to check the sign of `den`.
> 
> Ill-formed `Rational`s result in undefined behaviour.

The original inner constructor is now an outer constructor with the exact same specification as before. This PR is still entirely non-breaking (apart from the two bugfixes mentionned in the first message).