@Liozou 2020-04-17
Forgot to mention but the idea for `unsafe_rational` comes from @simeonschaub and thanks a lot to him for the thorough review!