@Liozou 2020-04-17
OK so the only CI error looks unrelated (it passes on https://github.com/JuliaLang/julia/commit/73c5a8238bc0232b775c89971c863cd62cd1b023 and I only fixed a docstring), and the PR is much better now with the `unsafe_rational` version. It only provides performance gain (and bugfixes) without breaking or exporting anything new so I guess it should be mergeable, but please let me know if you want me to change or fix something.