@JeffreySarnoff 2020-04-22
To facilitate evaluating this change, (using rational.jl) I created a module `Fractions.jl` that is this modification with the exported type `Fraction` rather than `Rational` and the exported infix op `⨸` (`\odiv`) rather than `\\`.  The module is given next.